separate Node & browser implementations
