+++ lib/mic-process.browser.js
@@ -1,0 +1,4 @@
+module.exports = function micProcess () {
+  // todo: use saebekassebil/microphone-stream or merge with ahdinosaur/read-audio
+  throw new Error('not implemented')
+}
